cascabel documentation wiki: cascabel manager source

+++ ccb/fcn/ccb_manager.h
@@ -12,11 +12,18 @@
 
 class CCBManager
 {
+
 public:
+
+	// construction
 	CCBManager(Frame* frame,Renderer2D* r2d,Camera2D* cam2d);
-	int add_lv(const char* path);
+
+	// functions
+	uint16_t add_lv(const char* path);
 	void dev_console(uint32_t &running,bool &dactive);
+
 private:
+
 	// background
 	Frame* m_frame;
 	Renderer2D* m_r2d;
